prepare for apps child PR
