+++ tools/block-generator/generator/generate.go
@@ -24,21 +24,23 @@ import (
 	"os"
 	"time"
 
+	"github.com/algorand/go-algorand/agreement"
 	cconfig "github.com/algorand/go-algorand/config"
 	"github.com/algorand/go-algorand/ledger"
 	"github.com/algorand/go-algorand/ledger/ledgercore"
 	"github.com/algorand/go-algorand/logging"
 	"github.com/algorand/go-algorand/protocol"
+	"github.com/algorand/go-algorand/rpcs"
 
-	"github.com/algorand/go-algorand/agreement"
 	"github.com/algorand/go-algorand/crypto"
 	"github.com/algorand/go-algorand/daemon/algod/api/server/v2/generated/model"
 	"github.com/algorand/go-algorand/data/basics"
 	"github.com/algorand/go-algorand/data/bookkeeping"
 	"github.com/algorand/go-algorand/data/committee"
 	"github.com/algorand/go-algorand/data/transactions"
-	"github.com/algorand/go-algorand/rpcs"
 )
+
+// ---- constructors ----
 
 // MakeGenerator initializes the Generator object.
 func MakeGenerator(dbround uint64, bkGenesis bookkeeping.Genesis, config GenerationConfig) (Generator, error) {
@@ -162,130 +164,53 @@ func MakeGenerator(dbround uint64, bkGenesis bookkeeping.Genesis, config Generat
 	return gen, nil
 }
 
-// Generator is the interface needed to generate blocks.
-type Generator interface {
-	WriteReport(output io.Writer) error
-	WriteGenesis(output io.Writer) error
-	WriteBlock(output io.Writer, round uint64) error
-	WriteAccount(output io.Writer, accountString string) error
-	WriteStatus(output io.Writer) error
-	WriteDeltas(output io.Writer, round uint64) error
-	Accounts() <-chan basics.Address
-	Stop()
+// initializeAccounting creates the genesis accounts.
+func (g *generator) initializeAccounting() {
+	if g.config.NumGenesisAccounts == 0 {
+		panic("Number of genesis accounts must be > 0.")
+	}
+
+	g.numPayments = 0
+	g.numAccounts = g.config.NumGenesisAccounts
+	for i := uint64(0); i < g.config.NumGenesisAccounts; i++ {
+		g.balances = append(g.balances, g.config.GenesisAccountInitialBalance)
+	}
 }
 
-type generator struct {
-	config GenerationConfig
-
-	// payment transaction metadata
-	numPayments uint64
-
-	// Number of algorand accounts
-	numAccounts uint64
-
-	// Block stuff
-	round         uint64
-	txnCounter    uint64
-	prevBlockHash string
-	timestamp     int64
-	protocol      protocol.ConsensusVersion
-	params        cconfig.ConsensusParams
-	genesis       bookkeeping.Genesis
-	genesisID     string
-	genesisHash   crypto.Digest
-
-	// Rewards stuff
-	feeSink                   basics.Address
-	rewardsPool               basics.Address
-	rewardsLevel              uint64
-	rewardsResidue            uint64
-	rewardsRate               uint64
-	rewardsRecalculationRound uint64
-
-	// balances for all accounts. To avoid crypto and reduce storage, accounts are faked.
-	// The account is based on the index into the balances array.
-	balances []uint64
-
-	// assets is a minimal representation of the asset holdings, it doesn't
-	// include the frozen state.
-	assets []*assetData
-	// pendingAssets is used to hold newly created assets so that they are not used before
-	// being created.
-	pendingAssets []*assetData
-
-	// apps is a minimal representation of the app holdings
-	apps map[appKind][]*appData
-	// pendingApps is used to hold newly created apps so that they are not used before
-	// being created.
-	pendingApps map[appKind][]*appData
-
-	transactionWeights []float32
-
-	payTxWeights   []float32
-	assetTxWeights []float32
-	appTxWeights   []float32
-
-	// Reporting information from transaction type to data
-	reportData Report
-
-	// ledger
-	ledger *ledger.Ledger
-
-	// cache the latest written block
-	latestBlockMsgp []byte
-
-	roundOffset uint64
+func (g *generator) initializeLedger() {
+	genBal := convertToGenesisBalances(g.balances)
+	// add rewards pool with min balance
+	genBal[g.rewardsPool] = basics.AccountData{
+		MicroAlgos: basics.MicroAlgos{Raw: g.params.MinBalance},
+	}
+	bal := bookkeeping.MakeGenesisBalances(genBal, g.feeSink, g.rewardsPool)
+	block, err := bookkeeping.MakeGenesisBlock(g.protocol, bal, g.genesisID, g.genesisHash)
+	if err != nil {
+		fmt.Printf("error making genesis: %v\n.", err)
+		os.Exit(1)
+	}
+	var prefix string
+	if g.genesisID == "" {
+		prefix = "block-generator"
+	} else {
+		prefix = g.genesisID
+	}
+	l, err := ledger.OpenLedger(logging.Base(), prefix, true, ledgercore.InitState{
+		Block:       block,
+		Accounts:    bal.Balances,
+		GenesisHash: g.genesisHash,
+	}, cconfig.GetDefaultLocal())
+	if err != nil {
+		fmt.Printf("error initializing ledger: %v\n.", err)
+		os.Exit(1)
+	}
+	g.ledger = l
 }
 
-type assetData struct {
-	assetID uint64
-	creator uint64
-	name    string
-	// Holding at index 0 is the creator.
-	holdings []*assetHolding
-	// Set of holders in the holdings array for easy reference.
-	holders map[uint64]*assetHolding
-}
-
-type appData struct {
-	appID   uint64
-	creator uint64
-	// TODO: more data, not sure yet exactly what
-}
-
-type assetHolding struct {
-	acctIndex uint64
-	balance   uint64
-}
-
-// Report is the generation report.
-type Report map[TxTypeID]TxData
-
-// TxData is the generator report data.
-type TxData struct {
-	GenerationTime  time.Duration `json:"generation_time_milli"`
-	GenerationCount uint64        `json:"num_generated"`
-}
-
-func track(id TxTypeID) (TxTypeID, time.Time) {
-	return id, time.Now()
-}
-func (g *generator) recordData(id TxTypeID, start time.Time) {
-	data := g.reportData[id]
-	data.GenerationCount++
-	data.GenerationTime += time.Since(start)
-	g.reportData[id] = data
-}
+// ---- implement Generator interface ----
 
 func (g *generator) WriteReport(output io.Writer) error {
 	return json.NewEncoder(output).Encode(g.reportData)
-}
-
-func (g *generator) WriteStatus(output io.Writer) error {
-	response := model.NodeStatusResponse{
-		LastRound: g.round + g.roundOffset,
-	}
-	return json.NewEncoder(output).Encode(response)
 }
 
 // WriteGenesis writes the genesis file and advances the round.
@@ -332,46 +257,6 @@ func (g *generator) WriteGenesis(output io.Writer) error {
 
 	_, err := output.Write(protocol.EncodeJSON(gen))
 	return err
-}
-
-func getTransactionOptions() []interface{} {
-	return []interface{}{paymentTx, assetTx}
-}
-
-func (g *generator) generateTransaction(round uint64, intra uint64) (transactions.SignedTxn, transactions.ApplyData, error) {
-	selection, err := weightedSelection(g.transactionWeights, getTransactionOptions(), paymentTx)
-	if err != nil {
-		return transactions.SignedTxn{}, transactions.ApplyData{}, err
-	}
-
-	switch selection {
-	case paymentTx:
-		return g.generatePaymentTxn(round, intra)
-	case assetTx:
-		return g.generateAssetTxn(round, intra)
-	default:
-		return transactions.SignedTxn{}, transactions.ApplyData{}, fmt.Errorf("no generator available for %s", selection)
-	}
-}
-
-func (g *generator) txnForRound(round uint64) uint64 {
-	// There are no transactions in the 0th round
-	if round == 0 {
-		return 0
-	}
-	return g.config.TxnPerBlock
-}
-
-// finishRound tells the generator it can apply any pending state.
-func (g *generator) finishRound(txnCount uint64) {
-	g.txnCounter += txnCount
-
-	g.timestamp += consensusTimeMilli
-	g.round++
-
-	// Apply pending assets...
-	g.assets = append(g.assets, g.pendingAssets...)
-	g.pendingAssets = nil
 }
 
 // WriteBlock generates a block full of new transactions and writes it to the writer.
@@ -493,6 +378,72 @@ func (g *generator) WriteBlock(output io.Writer, round uint64) error {
 	return nil
 }
 
+func (g *generator) WriteAccount(output io.Writer, accountString string) error {
+	addr, err := basics.UnmarshalChecksumAddress(accountString)
+	if err != nil {
+		return fmt.Errorf("failed to unmarshal address: %w", err)
+	}
+
+	idx := accountToIndex(addr)
+
+	// Asset Holdings
+	assets := make([]model.AssetHolding, 0)
+	createdAssets := make([]model.Asset, 0)
+	for _, a := range g.assets {
+		// holdings
+		if holding := a.holders[idx]; holding != nil {
+			assets = append(assets, model.AssetHolding{
+				Amount:   holding.balance,
+				AssetID:  a.assetID,
+				IsFrozen: false,
+			})
+		}
+		// creator
+		if len(a.holdings) > 0 && a.holdings[0].acctIndex == idx {
+			nameBytes := []byte(a.name)
+			asset := model.Asset{
+				Index: a.assetID,
+				Params: model.AssetParams{
+					Creator:  accountString,
+					Decimals: 0,
+					Clawback: &accountString,
+					Freeze:   &accountString,
+					Manager:  &accountString,
+					Reserve:  &accountString,
+					Name:     &a.name,
+					NameB64:  &nameBytes,
+					Total:    assetTotal,
+				},
+			}
+			asset.Params.DefaultFrozen = new(bool)
+			*(asset.Params.DefaultFrozen) = false
+			createdAssets = append(createdAssets, asset)
+		}
+	}
+
+	data := model.Account{
+		Address:                     accountString,
+		Amount:                      g.balances[idx],
+		AmountWithoutPendingRewards: g.balances[idx],
+		AppsLocalState:              nil,
+		AppsTotalExtraPages:         nil,
+		AppsTotalSchema:             nil,
+		Assets:                      &assets,
+		AuthAddr:                    nil,
+		CreatedApps:                 nil,
+		CreatedAssets:               &createdAssets,
+		Participation:               nil,
+		PendingRewards:              0,
+		RewardBase:                  nil,
+		Rewards:                     0,
+		Round:                       g.round - 1,
+		SigType:                     nil,
+		Status:                      "Offline",
+	}
+
+	return json.NewEncoder(output).Encode(data)
+}
+
 // WriteDeltas generates returns the deltas for payset.
 func (g *generator) WriteDeltas(output io.Writer, round uint64) error {
 	// the first generated round has no statedelta.
@@ -520,37 +471,70 @@ func (g *generator) WriteDeltas(output io.Writer, round uint64) error {
 	return nil
 }
 
-// initializeAccounting creates the genesis accounts.
-func (g *generator) initializeAccounting() {
-	if g.config.NumGenesisAccounts == 0 {
-		panic("Number of genesis accounts must be > 0.")
+func (g *generator) WriteStatus(output io.Writer) error {
+	response := model.NodeStatusResponse{
+		LastRound: g.round + g.roundOffset,
 	}
-
-	g.numPayments = 0
-	g.numAccounts = g.config.NumGenesisAccounts
-	for i := uint64(0); i < g.config.NumGenesisAccounts; i++ {
-		g.balances = append(g.balances, g.config.GenesisAccountInitialBalance)
-	}
+	return json.NewEncoder(output).Encode(response)
 }
 
-func signTxn(txn transactions.Transaction) transactions.SignedTxn {
-	stxn := transactions.SignedTxn{
-		Sig:      crypto.Signature{},
-		Msig:     crypto.MultisigSig{},
-		Lsig:     transactions.LogicSig{},
-		Txn:      txn,
-		AuthAddr: basics.Address{},
-	}
+// Stop cleans up allocated resources.
+func (g *generator) Stop() {
+	g.ledger.Close()
+}
 
-	// TODO: Would it be useful to generate a random signature?
-	stxn.Sig[32] = 50
+// Accounts is used in the runner to generate a list of addresses.
+func (g *generator) Accounts() <-chan basics.Address {
+	results := make(chan basics.Address, 10)
+	go func() {
+		defer close(results)
+		for i := uint64(0); i < g.numAccounts; i++ {
+			results <- indexToAccount(i)
+		}
+	}()
+	return results
+}
 
-	return stxn
+// ---- transaction options vectors ----
+
+func getTransactionOptions() []interface{} {
+	return []interface{}{paymentTx, assetTx}
 }
 
 func getPaymentTxOptions() []interface{} {
 	return []interface{}{paymentTx, paymentAcctCreateTx}
 }
+
+func getAssetTxOptions() []interface{} {
+	return []interface{}{assetCreate, assetDestroy, assetOptin, assetXfer, assetClose}
+}
+
+func getAppTxOptions() []interface{} {
+	return []interface{}{
+		appSwapCreate, appSwapUpdate, appSwapDelete, appSwapOptin, appSwapCall, appSwapClose, appSwapClear,
+		appBoxesCreate, appBoxesUpdate, appBoxesDelete, appBoxesOptin, appBoxesCall, appBoxesClose, appBoxesClear,
+	}
+}
+
+// ---- Transaction Generation (Pay/Asset/Apps) ----
+
+func (g *generator) generateTransaction(round uint64, intra uint64) (transactions.SignedTxn, transactions.ApplyData, error) {
+	selection, err := weightedSelection(g.transactionWeights, getTransactionOptions(), paymentTx)
+	if err != nil {
+		return transactions.SignedTxn{}, transactions.ApplyData{}, err
+	}
+
+	switch selection {
+	case paymentTx:
+		return g.generatePaymentTxn(round, intra)
+	case assetTx:
+		return g.generateAssetTxn(round, intra)
+	default:
+		return transactions.SignedTxn{}, transactions.ApplyData{}, fmt.Errorf("no generator available for %s", selection)
+	}
+}
+
+// ---- 1. Pay Transactions ----
 
 // generatePaymentTxn creates a new payment transaction. The sender is always a genesis account, the receiver is random,
 // or a new account.
@@ -602,15 +586,24 @@ func (g *generator) generatePaymentTxnInternal(selection TxTypeID, round uint64,
 	return signTxn(txn), transactions.ApplyData{}, nil
 }
 
-func getAssetTxOptions() []interface{} {
-	return []interface{}{assetCreate, assetDestroy, assetOptin, assetXfer, assetClose}
-}
+// ---- 2. Asset Transactions ----
 
-func getAppTxOptions() []interface{} {
-	return []interface{}{
-		appSwapCreate, appSwapUpdate, appSwapDelete, appSwapOptin, appSwapCall, appSwapClose, appSwapClear,
-		appBoxesCreate, appBoxesUpdate, appBoxesDelete, appBoxesOptin, appBoxesCall, appBoxesClose, appBoxesClear,
+func (g *generator) generateAssetTxn(round uint64, intra uint64) (transactions.SignedTxn, transactions.ApplyData, error) {
+	start := time.Now()
+	selection, err := weightedSelection(g.assetTxWeights, getAssetTxOptions(), assetXfer)
+	if err != nil {
+		return transactions.SignedTxn{}, transactions.ApplyData{}, err
 	}
+
+	actual, txn := g.generateAssetTxnInternal(selection.(TxTypeID), round, intra)
+	defer g.recordData(actual, start)
+
+	if txn.Type == "" {
+		fmt.Println("Empty asset transaction.")
+		os.Exit(1)
+	}
+
+	return signTxn(txn), transactions.ApplyData{}, nil
 }
 
 func (g *generator) generateAssetTxnInternal(txType TxTypeID, round uint64, intra uint64) (actual TxTypeID, txn transactions.Transaction) {
@@ -766,133 +759,50 @@ func (g *generator) generateAssetTxnInternalHint(txType TxTypeID, round uint64, 
 	return
 }
 
-func (g *generator) generateAssetTxn(round uint64, intra uint64) (transactions.SignedTxn, transactions.ApplyData, error) {
-	start := time.Now()
-	selection, err := weightedSelection(g.assetTxWeights, getAssetTxOptions(), assetXfer)
-	if err != nil {
-		return transactions.SignedTxn{}, transactions.ApplyData{}, err
-	}
+// ---- miscellaneous ----
 
-	actual, txn := g.generateAssetTxnInternal(selection.(TxTypeID), round, intra)
-	defer g.recordData(actual, start)
-
-	if txn.Type == "" {
-		fmt.Println("Empty asset transaction.")
-		os.Exit(1)
-	}
-
-	return signTxn(txn), transactions.ApplyData{}, nil
+func track(id TxTypeID) (TxTypeID, time.Time) {
+	return id, time.Now()
 }
 
-func (g *generator) initializeLedger() {
-	genBal := convertToGenesisBalances(g.balances)
-	// add rewards pool with min balance
-	genBal[g.rewardsPool] = basics.AccountData{
-		MicroAlgos: basics.MicroAlgos{Raw: g.params.MinBalance},
-	}
-	bal := bookkeeping.MakeGenesisBalances(genBal, g.feeSink, g.rewardsPool)
-	block, err := bookkeeping.MakeGenesisBlock(g.protocol, bal, g.genesisID, g.genesisHash)
-	if err != nil {
-		fmt.Printf("error making genesis: %v\n.", err)
-		os.Exit(1)
-	}
-	var prefix string
-	if g.genesisID == "" {
-		prefix = "block-generator"
-	} else {
-		prefix = g.genesisID
-	}
-	l, err := ledger.OpenLedger(logging.Base(), prefix, true, ledgercore.InitState{
-		Block:       block,
-		Accounts:    bal.Balances,
-		GenesisHash: g.genesisHash,
-	}, cconfig.GetDefaultLocal())
-	if err != nil {
-		fmt.Printf("error initializing ledger: %v\n.", err)
-		os.Exit(1)
-	}
-	g.ledger = l
+func (g *generator) recordData(id TxTypeID, start time.Time) {
+	data := g.reportData[id]
+	data.GenerationCount++
+	data.GenerationTime += time.Since(start)
+	g.reportData[id] = data
 }
 
-// Stop cleans up allocated resources.
-func (g *generator) Stop() {
-	g.ledger.Close()
+func (g *generator) txnForRound(round uint64) uint64 {
+	// There are no transactions in the 0th round
+	if round == 0 {
+		return 0
+	}
+	return g.config.TxnPerBlock
 }
 
-func (g *generator) WriteAccount(output io.Writer, accountString string) error {
-	addr, err := basics.UnmarshalChecksumAddress(accountString)
-	if err != nil {
-		return fmt.Errorf("failed to unmarshal address: %w", err)
-	}
+// finishRound tells the generator it can apply any pending state.
+func (g *generator) finishRound(txnCount uint64) {
+	g.txnCounter += txnCount
 
-	idx := accountToIndex(addr)
+	g.timestamp += consensusTimeMilli
+	g.round++
 
-	// Asset Holdings
-	assets := make([]model.AssetHolding, 0)
-	createdAssets := make([]model.Asset, 0)
-	for _, a := range g.assets {
-		// holdings
-		if holding := a.holders[idx]; holding != nil {
-			assets = append(assets, model.AssetHolding{
-				Amount:   holding.balance,
-				AssetID:  a.assetID,
-				IsFrozen: false,
-			})
-		}
-		// creator
-		if len(a.holdings) > 0 && a.holdings[0].acctIndex == idx {
-			nameBytes := []byte(a.name)
-			asset := model.Asset{
-				Index: a.assetID,
-				Params: model.AssetParams{
-					Creator:  accountString,
-					Decimals: 0,
-					Clawback: &accountString,
-					Freeze:   &accountString,
-					Manager:  &accountString,
-					Reserve:  &accountString,
-					Name:     &a.name,
-					NameB64:  &nameBytes,
-					Total:    assetTotal,
-				},
-			}
-			asset.Params.DefaultFrozen = new(bool)
-			*(asset.Params.DefaultFrozen) = false
-			createdAssets = append(createdAssets, asset)
-		}
-	}
-
-	data := model.Account{
-		Address:                     accountString,
-		Amount:                      g.balances[idx],
-		AmountWithoutPendingRewards: g.balances[idx],
-		AppsLocalState:              nil,
-		AppsTotalExtraPages:         nil,
-		AppsTotalSchema:             nil,
-		Assets:                      &assets,
-		AuthAddr:                    nil,
-		CreatedApps:                 nil,
-		CreatedAssets:               &createdAssets,
-		Participation:               nil,
-		PendingRewards:              0,
-		RewardBase:                  nil,
-		Rewards:                     0,
-		Round:                       g.round - 1,
-		SigType:                     nil,
-		Status:                      "Offline",
-	}
-
-	return json.NewEncoder(output).Encode(data)
+	// Apply pending assets...
+	g.assets = append(g.assets, g.pendingAssets...)
+	g.pendingAssets = nil
 }
 
-// Accounts is used in the runner to generate a list of addresses.
-func (g *generator) Accounts() <-chan basics.Address {
-	results := make(chan basics.Address, 10)
-	go func() {
-		defer close(results)
-		for i := uint64(0); i < g.numAccounts; i++ {
-			results <- indexToAccount(i)
-		}
-	}()
-	return results
+func signTxn(txn transactions.Transaction) transactions.SignedTxn {
+	stxn := transactions.SignedTxn{
+		Sig:      crypto.Signature{},
+		Msig:     crypto.MultisigSig{},
+		Lsig:     transactions.LogicSig{},
+		Txn:      txn,
+		AuthAddr: basics.Address{},
+	}
+
+	// TODO: Would it be useful to generate a random signature?
+	stxn.Sig[32] = 50
+
+	return stxn
 }
